third_party/WebKit/LayoutTests/editing/deleting/pruning_after_merge.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>',
                'b^oo<b><div>b|ar</div><div>baz</div></b>',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable>',
                'b|<b></b>',
                '<b><div style="display: inline !important;">ar</div></b>',
                '<b><div>baz</div></b>',
            '</div>',
        ].join('')),
        'Merge multiple blocks with display:inline');

    test(() => assert_selection(
        [
            '<div contenteditable>',
                '<div>foo</div>',
                '<div style="font-weight: bold;">',
                    '<div>|bar</div>',
                '</div>',
            '</div>',
        ].join(''),
        'delete',
        [
            '<div contenteditable>',
                '<div>',
                    'foo|<span style="font-weight: bold;">bar</span>',
                '</div>',
            '</div>',
        ].join('')),
        'Delete DIV with STYLE to SPAN with STYLE');
</script>
